References

UniformTEMPO.UniformPTMPO — Type

UnfiromPTMPO(s_dim::Int, delta_t::Real, q::AbstractArray{<:Number,4}, v_r::AbstractVector{<:Number}, v_l::Transpose{<:Number,<:AbstractVector})
UniformPTMPO(s_dim::Int, delta_t::Real)

a struct representing a Uniform process tensor MPO. The second constructor generates a trivial process tensor.

Fields

s_dim system dimension

delta_t Trotter time-step

q uniform tensor of the UniformPTMPO (system-bath propagator). Dimensions are (aux-out, sys-out, aux-in, sys-in)

v_r right boundary vector (bath initial state)

v_l left boundary vector (bath trace). Must be transpose type.

Related Functions

bond_dim to return the bond dimension.

+ for adding two process tensors.

* for combining a process tensor with a local channel.

^ for taking powers (changing the time-step).

steadystate to compute the steady state.

evolve for density matrix time evolution.

channel and choi_channel to compute quantum channels.

process_tensor to compute process tensors.

two_point_correlations, two_point_correlations_freq and susceptibility to compute unequal time correlation functions.

spectrum to compute the transfer matrix spectrum.

truncate to truncate the bond dimension of the process tensor.

Base.:+ — Method

+(a::UniformPTMPO, b::UniformPTMPO)

combine two Uniform PT-MPOs a and b in second order Trotter using the ACE combination scheme [Cygorek et.al., Nature Physics 18, 6 662-668 (2022)].

Base.:* — Method

*(pt::UniformPTMPO, Φ::AbstractMatrix{<:Number})

Combine a process tensor with a channel Φ on the right.

Base.:* — Method

*(Φ::AbstractMatrix{<:Number}, pt::UniformPTMPO)

Combine a process tensor with a channel Φ on the left.

Base.:^ — Method

^(pt::UniformPTMPO, y::Real)

Take the power of a uniform process tensor to change the time step.

UniformTEMPO.include_system_hamiltonian — Function

include_system_hamiltonian(pt::UniformPTMPO, h_s::AbstractMatrix{<:Number})

Include an additional system Hamiltonian h_s into the process tensor pt via Trotterization.

UniformTEMPO.bond_dim — Function

bond_dim(pt::UniformPTMPO)

Returns the bond dimension of a process tensor.

UniformTEMPO.uniTEMPO — Function

uniTEMPO(s::Union{AbstractMatrix{<:Number}, Vector}, delta_t::Real, bcf::Union{Function, Array}, tol::Real; auto_nc::Bool=true, n_c::Int=100_000, truncation::Symbol=:rel, cap_rank::Int=100_000, ftype::DataType=Float64, low_rank_svd::Bool=false, svd_filtering_tol::Real=0, max_rank::Int=100_000)

Compute the process tensor for a Gaussian bath using the uniTEMPO algorithm. Returns a UniformPTMPO representing the compressed process tensor.

Arguments

s is the coupling operator (must be hermitian) or a list of coupling operators.

delta_t is the Trotter time step.

bcf is the bath correlation function.

tol is the SVD compression tolerance, determining the accuracy of the compression.

Keyword Arguments

auto_nc if set to true, a memory cutoff is determined automatically based on the decay of the bath correlation function and the provided tolerance (recommended). Otherwise, the user can specify n_c directly.

n_c is the maximum memory cutoff. Note that auto_nc will not search beyond this value.

truncation specifies the truncation scheme for singular value truncation. It can be :rel for relative, :abs for absolute, see function new_rank.

cap_rank fixes the maximum bond dimension allowed during the SVD compression. Usually it is not recommended to use this option.

max_rank is a hard limit on the bond dimension. The algorithm will throw an error if this limit is reached.

low_rank_svd if set to true, compute the low-rank subspace first and then perform SVD on the low-rank subspace only. This can be beneficial for simulations with large system dimensions. It is recommended to set truncation to :abs when using this option to avoid issues with determining the rank based on the relative cutoff when the full singular value spectrum is not computed.

svd_filtering_tol if set to a nonzero value, use an iTEBD scheme with SVD filtering on the physical indices. The extra filtering is performed with the given tolerance.

UniformTEMPO.evolve — Function

evolve(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, n::Int; h_s::Union{Nothing, AbstractMatrix{<:Number}, Function}=nothing, return_full::Bool=false)

Compute the time evolution for n time steps with initial state ρ_0 (density matrix). h_s can be an additional systm Hamiltonian (may also be a time-dependent Hamiltonian).

The function simply returns a vector of density matrices at each time step.

Example

ρ_t = evolve(pt, n, ρ_0; h_s=h_s)
t_eval = collect(0:n) * pt.delta_t
plot(t_eval, real.([tr(O * ρ_t[i]) for i in eachindex(ρ_t)])) # plot the dynamics of ⟨O⟩ = tr(O*rho)

UniformTEMPO.channel — Function

channel(pt::UniformPTMPO, n::Int; h_s::Union{Nothing,AbstractMatrix{<:Number},Function}=nothing)

compute dynamical map for n time-steps in generic basis (linear map on density matrix). h_s is an optional additional system Hamiltonian. The map can be applied as follows

Example

ρ_t = evolve(pt, ρ_0, n)
Φ_t = channel(pt, n)
ρ̃_t = [reshape(Φ_t[i] * ρ_0[:], size(ρ_0)) for i in eachindex(Φ_t)]
ρ_t ≈ ρ̃_t # true

UniformTEMPO.choi_channel — Function

choi_channel(pt::UniformPTMPO, n::Int; h_s::Union{Nothing,AbstractMatrix{<:Number},Function}=nothing)

compute the dynamical map for given hamiltonian h_s for n time-steps in choi representation.

UniformTEMPO.steadystate — Function

steadystate(pt::UniformPTMPO; h_s::AbstractMatrix{<:Number}=zeros(pt.s_dim, pt.s_dim), return_full::Bool=false, ED::Bool=false)

Compute the steady state of a process tensor pt. An additional system Hamiltonian h_s can be specified directly for convenience.

Returns the system steady state if return_full=false and the full leading eigenvector of q if return_full=true. Set ED=true to use exact diagonalization instead of an iterative eigensolver.

UniformTEMPO.spectrum — Function

spectrum(pt::UniformPTMPO; h_s::AbstractMatrix{<:Number}=zeros(pt.s_dim, pt.s_dim); enforce_physical=false)

Compute the spectrum of the process tensor transfer matrix for Hamiltonian h_s using full diagonalization.

When setting enforce_physical=true the dominant decay rate is truncated to zero explicitly.

Returns the spectrum γ and the basis change matrix R such that q(t)=R*exp.(diagm(γ)*t)*inv(R), where q(t) is the propagator for time t.

UniformTEMPO.two_point_correlations — Function

two_point_correlations(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, n_1::Int, n_2::Int, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}; h_s::Union{Nothing,AbstractMatrix{<:Number},Function} = nothing)

Calculate the two-time correlation function $⟨o₂(t+s)o₁(s)⟩$ for operators o_1, o_2. h_s is the system Hamiltonian, n_1 is the number of initial time steps where s = pt.delta_t * n_1 and n_2 the number of evolution time steps t = pt.delta_t * n_2. Returns the correlation function for all n <= n_2. By default, the operators are left acting (Keldysh "+" contour). For right acting operators directly provide the superoperators.

two_point_correlations(pt::UniformPTMPO, n::Int, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}; h_s::Union{Nothing,AbstractMatrix{<:Number},Function} = nothing)

Calculate the stationary two-time correlation function $⟨o₂(t)o₁(0)⟩$ for operators o_1, o_2. h_s is the system Hamiltonian and n is the number of time steps (as in evolve).

UniformTEMPO.two_point_correlations_fourier — Function

two_point_correlations_fourier(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, ω::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

Directly compute the half-sided fourier transform of the stationary two point function $c(t) = ⟨o_2(t)o_1(0)⟩$ evalutated at frequencies ω.

two_point_correlations_fourier(pt::UniformPTMPO, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, ω::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

If no initial state is passed then this computes the stationary correlation function.

This function accepts both superoperators and normal operators. Normal operators are always interpreted as left-acting. For right-acting operators, provide the corresponding superoperator o_right = kron(transpose(o), I(size(o, 1))).

UniformTEMPO.three_point_correlations_fourier — Function

three_point_correlations_fourier(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, o_3::AbstractMatrix{<:Number}, ω_12::AbstractVector{<:Real}, ω_23::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

Directly compute the half-sided fourier transform of the three point function R(t, s) = ⟨o_3(t)o_2(s)o_1(0)⟩ evalutated at frequencies ω_12 (time s) and ω_23 (time t).

three_point_correlations_fourier(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, o_3::AbstractMatrix{<:Number}, ω_12::AbstractVector{<:Real}, ω_23::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

Stationary version that uses the steadystate as initial state.

UniformTEMPO.four_point_correlations_fourier — Function

four_point_correlations_fourier(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, o_3::AbstractMatrix{<:Number}, o_4::AbstractMatrix{<:Number}, ω_12::AbstractVector{<:Real}, n_23::Int, ω_34::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

Directly compute the half-sided fourier transform of the four point correlation function R(t, τ) = ⟨o4(t)o3(τ+T)o2(τ)o1(0)⟩ evalutated at frequencies ω12 (τ) and ω34 (t), with waiting time T (n_23 time-steps).

four_point_correlations_fourier(pt::UniformPTMPO, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, o_3::AbstractMatrix{<:Number}, o_4::AbstractMatrix{<:Number}, ω_12::AbstractVector{<:Real}, n_23::Int, ω_34::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

Stationary version that uses the steadystate as initial state.

UniformTEMPO.susceptibility — Function

susceptibility(pt::UniformPTMPO, o_1::AbstractMatrix{<:Number}, o_2::AbstractMatrix{<:Number}, ω::AbstractVector{<:Real}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

Directly compute the fourier transform of the stationary susceptibility $χ(t-s) = iΘ(t-s) ⟨[o_2(t),o_1(s)]⟩$ evalutated at frequencies ω.

UniformTEMPO.process_tensor — Function

process_tensor(pt::UniformPTMPO, ρ_0::AbstractMatrix{<:Number}, n::AbstractVector{<:Int}; h_s::Union{Nothing,AbstractMatrix{<:Number}}=nothing)

compute the process tensor in canonical (Liouville-space) basis for interventions at time steps n. Returns the process tensor in the form

\[T[(μ^1_I,μ^1_O),(μ^2_I,μ^2_O),...,μ^N_I]\]

where N=length(n) and $μ^i_{I/0}=1...d^2$. The trivial output index for the last time step is omitted. If no initial state is provided, the output index for the initial state is included as a first index. To convert a process tensor of this form to choi representation, use pt_tensor_to_choi. See [Pollock et.al., Phys. Rev. A 97, 012127 (2018)] for details.

UniformTEMPO.pt_to_choi — Function

pt_to_choi(pt::AbstractArray{<:Number})

Convert a process tensor (from the process_tensor function) to a Choi state. Returns the choi state in the form

\[ρ_{choi} = Σ_{i,i'} Y[i_0,i_1,...,i_N;i_0',i_1',...,i_N'] |i_0,i_1,...,i_N⟩ ⟨i_0',i_1',...,i_N'|\]

where N in the number of iterventions. i_0 corresponds to th^e input state (if applicable), see also the docstring for process_tensor. For details on the Choi mapping see [Pollock et.al., Phys. Rev. A 97, 012127 (2018)].

UniformTEMPO.floquet_process_tensor — Function

floquet_process_tensor(pt::UniformPTMPO, h_s::Function, Δt::Real)

Construct a Floquet process tensor from a uniform PT-MPO and a periodic system Hamiltonian h_s. The period is Δt.

UniformTEMPO.truncate — Function

truncate(pt::UniformPTMPO, tol::Real; pre_svd::Bool=true)

truncate the bond dimension of a UniformPTMPO based on a tolerance tol.

UniformTEMPO.to_ITensor — Function

to_ITensor(pt::UniformPTMPO; tags=[])

Convert a UniformPTMPO to a UniformPTMPOITensor where the data is stored as ITensors. Load ITensors.jl via using ITensors for this function to work.